// src/model/Company.ts
export class Company {
  idCompany: any;
  nameCompany: any;
	email: any;
	country: any;
  contactPerson:any;
  positionInCompany:any;
  city: any;
	street: any;
  postalCode:any;
  webAdress:any;
	logo: any;
	suffix:any;
	phone: any;
	fax: any;
  yearEstablished:any;
  status:any;
  type:any;
  listReferences:any;
  createdTime:any;
  verificationCode:any;
	activity:any;
	enabled: any;
	language: any;
  deleted: any;
}