// src/model/Profile.ts
export class Profile{
   /**
  id: number;
  salutation : string;
  nom: string;
  prenom: string;
	Email: string;
	birthdate: Date;
	country: string;
  city: string;
	address: string;
	photo: string;
	suffix:number;
	phone: number;
	phone2: number;
	bio:string;
	profession:string;
	nationality:string;
	activity:string;
	deleted: boolean;
	enabled: boolean;
	language: string;
  */
  idProfile: any;
  salutation: any;
  nom: any;
  prenom: any;
	email: any;
  createdTime:any;
	birthdate: any;
	country: any;
  city: any;
	address: any;
	photo: any;
	suffix:any;
	phone: any;
	phone2: any;
	bio:any;
	profession:any;
	nationality:any;
	activity:any;
	deleted: any;
	enabled: any;
	language: any;
}